Per Gary Granger's suggestion I rewrote signal handling according to these guidelines: http://doc.qt.nokia.com/4.6/unix-signals.html
I've used a single slot to handle all types of SIG events.

The UNIX signal handler also throws an IOException so that it can break the point at
which the Calibrator::setup's call to SampleInputStream::readInputHeader is hung.

This hung situation is normal and it occurs when dsm_server is running and no DSM's
are connected to it yet.