Modify signal handler to reset all A2D channels
This program occasionally hangs in mid-calibration.  User is forced to
ctrl-C program.  This leaves some channels receiving a fixed voltage.
Have the signal handler call acc->SetNextCalVoltage(DONE).  This should
reset all the channels.

Since sig handlers are static methods, had to add an instance pointer.
This class is a singleton, so should be good.

This has not been tested yet.  Will do so...